Use obj.data() instead of &obj.at(0) to prevent exception
In case of a 0-sized container, &obj.at(0) would throw an exception
even if we don't intend to use the pointer afterward.
obj.data is required to provide a non-nullptr pointer that we should
not dereference for most containers (except strings since C++11),
but that's fine for the way we intend (not to) use it.

(cherry picked from commit c2a16005ce822b3d704d2e717d734ee94e4ed189)