dnsdist: Detect ECS collisions in the packet cache
We did not detect a collision involving two queries for the same qname,
qtype, qclass and flags but with different ECS values hashing to the same key.
This commit adds an option to the packet cache to parse and keep the ECS value
in order to detect this kind of collisions.
It's not enabled by default because parsing the ECS value has a cost.

(cherry picked from commit f89b3bd8d75444957a1a4f3c6499fd9e09abbd47)