auth: In `Bind2Backend::lookup()`, use the `zoneId` when we have it
After the initial lookup corresponding to a `DNSBackend::getAuth()`,
the subsequent ones already have the `zoneId`, so use it instead of
looping on `chopOff()` again. This should be much more efficient.

Please note that I tried this optimization before and it caused an
issue because the Bind backend generates its IDs dynamically and
those are very likely to collide with the ones from a different
backend in multi-backends mode. This commit adds an additional
check that the qname does indeed belongs to the zone, as we do
in the SQL backends, so it should be fine, unless you host a
zone and a sub-zone in different backends and they end up sharing
the same ID.
If we want to make sure that doesn't happen, we need to keep track
of which backend returned the zone ID.