auth: Keep record content in native format in the bind backend
Keep the records in the 'native' DNSZoneRecord format in memory,
preventing a conversion from string to our internal format for
every lookup.
This consumes a bit more of memory (~20%), and slows down a bit some
operations like searching into records because they need to be
converted back to string format.