auth: Do an ANY lookup for all types then filter
Most of our backends have a very high latency, meaning that it takes
a long time to send a query and get the answer, regardless of whether
we are asking for one type or several.
Our code base often asks for a specific type, and our current code
stores separately the answers for ANY queries and the ones for a
specific type. This seems wasteful since the answer to an ANY query
already contains the records for a more specific one, and our
in-memory records cache is must faster than going to the backend.
We could save a round-trip by looking for ANY answers when we don't
find a specific one in the record cache, but our first query is often
for the specific NS type because we are looking for a referral.
This PR converts all lookups to an ANY lookup instead, making sure
that we fill the cache as fast as possible to save round-trips to
the backend later.
My tests showed roughly one third less queries to the backend in simple
cases, and probably more in DNSSEC cases, while achieving higher QPS
boundaries (~ +30%). CPU usage is also significantly reduced while
replying a real-world PCAP.
The number of entries in the records cache is also significantly
lower since we don't need to store a record twice, for ANY and
for the exact type itself.

We could easily enable that change for specific backends only if
we believe it might have a negative effect on some of them, although
testing with the bind backend showed a slight improvement there as well,
even though lookups in the bind backend are already quite fast.
I have not tested LMDB.

We could reduce the number of round-trips to the backend even more
by getting rid of the 'SOA' special case, since I'm not aware of
any backend currently implementing it in a special way.