auth: Speedup DNSRecordContent::deserialize() a little
The move from a vector to a PacketBuffer prevents the memory from
being zeroed while we are going to overwrite it right away.
The change from MOADNSParser to PacketReader avoids the need to
copy the header and the whole packet (allocation + copy),
parse the qname again, as well as the allocation of a vector of records
while we know we have only one of these.