Only use a shared pointer for DNSRecordContent when we need to
Shared pointers are quite expensive while unique pointers are cheap,
and converting a unique pointer to a shared pointer when we need to
is not more expensive than going with a shared pointer in the first
place, so let's not pay the cost up front.

The logical next step would be to stop allocating shared pointers
in MOADNSParser, and only convert the unique pointers to shared pointers
if needed.