auth: Speed up ECDSA and RSA signatures
For ECDSA, and likely for RSA, computing the public key is not a cheap
operation. So instead of computing it twice to get the lookup key for
our signatures cache, reuse the computed public key and only compute its
digest.
In my tests this reduces by 25% the cost of calling addRRSigs() for ECDSA
signatures when the signature is already present in the cache.